telegram-bot/requirements.txt
python-telegram-bot==20.7
flask==3.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
